refactor: move MaterialPropsGroup into its own pattern file
